Added restrictions to images
Added attribute "allowableResponses" which you pass a list of grid boxes into. It will then only allow you to click inside those boxes. If you click on a different one, it will not show up or get recorded.
If no allowableResponses is provided, it will assume all are allowed

Fixed Excel that had a rotated 180 field image, so the grids were wrong.